Refactor inset adjuster to separate file and better name.
